requirements.txt
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0
pytest-cov>=3.0.0
kubernetes>=25.0.0
boto3>=1.26.0
pyyaml>=6.0
requests>=2.28.0
urllib3>=1.26.0
paramiko>=2.11.0
jinja2>=3.1.0
python-dotenv>=1.0.0
tabulate>=0.9.0
colorama>=0.4.0
click>=8.1.0
